requirements-vercel.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
numpy==1.24.3
Pillow==10.0.1 